pyproject.toml
[tool.poetry]
name = "langgraph-projects"
version = "0.1.0"
description = "LangChain and LangGraph development environment"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.21"
langchain-core = "^0.3.51"
langchain-community = "^0.3.21"
langgraph = "^0.4.0"
langsmith = "^0.3.0"
python-dotenv = "^1.1.0"
gradio = "^5.0.0"
openai = "^1.72.0"
jupyterlab = "^4.4.0"
numpy = "^2.2.4"
pandas = "^2.2.3"
anthropic = "^0.52.0"
langchain-openai = "^0.3.12"
langchain-anthropic = "^0.3.10"
langchain-tavily = "^0.2.0"
ipython = "^9.1.0"
langchain-experimental = "^0.3.4"
ipykernel = "^6.29.5"
scikit-learn = "^1.7.0"
pyarrow = "^20.0.0"
langgraph-checkpoint-sqlite = "^2.0.10"
langgraph-sdk = "^0.1.70"
langgraph-cli = "^0.2.12"
langgraph-api = "^0.2.45"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^25.0.0"
isort = "^6.0.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"